implementing react router to allow for clickable links to view characters in a specific episode
